requirements.txt
kenlm
python-dotenv
pandas
pyyaml
plotly
tabulate
matplotlib
nltk
requests
symspellpy